UI updates, tools options and js modules
1. Add expect reply date display to message header view.
2. Add one preference entry, indicating wheather the user wants to
create calendar events. The button that toggles this preference is added
to the tools menu.
3. Add two javascript modules handling gloda queries and calendar
interaction.

Lightning is broken on my Daily build so I can't test the calendar
interaction code. But the code is directly imported from my extension
that was tested working properly so it should work fine.